% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanandMedianFunction.R
\name{MeanandMedian}
\alias{MeanandMedian}
\title{Summary Statistics: Mean and Median
The function created above returns the mean and median of a numeric variable to the user.}
\usage{
MeanandMedian(data)
}
\arguments{
\item{data}{is the input we will provide for the function to run on.}

\item{mean}{is the mean of the numeric variable.}

\item{median}{is the median of the numeric variable.}
}
\value{
the print command that returns the value of mean and median for the chosen numeric #' variable or if the input is not numeric, it will return an error as defined.
}
\description{
Summary Statistics: Mean and Median
The function created above returns the mean and median of a numeric variable to the user.
}
\examples{
MeanandMedian(c(12,54,23,78,33))
library(palmerpenguins)
MeanandMedian(penguins$bill_length_mm)
MeanandMedian(penguins$island)
}
